added a lock for the storages which can be activated as long as a device is tracking to avoid modifications

+++ Modules/IGT/DataManagement/mitkNavigationToolStorage.cpp
@@ -25,7 +25,7 @@ const std::string  mitk::NavigationToolStorage::US_INTERFACE_NAME = "org.mitk.se
 const std::string  mitk::NavigationToolStorage::US_PROPKEY_SOURCE_ID = US_INTERFACE_NAME + ".sourceID";
 const std::string  mitk::NavigationToolStorage::US_PROPKEY_STORAGE_NAME = US_INTERFACE_NAME + ".name";
 
-mitk::NavigationToolStorage::NavigationToolStorage() : m_ToolCollection(std::vector<mitk::NavigationTool::Pointer>()),m_DataStorage(NULL)
+mitk::NavigationToolStorage::NavigationToolStorage() : m_ToolCollection(std::vector<mitk::NavigationTool::Pointer>()),m_DataStorage(NULL),m_storageLocked(false)
   {
   this->SetName("ToolStorage (no name given)");
   }
@@ -74,7 +74,17 @@ void mitk::NavigationToolStorage::UnRegisterMicroservice(){
 
 bool mitk::NavigationToolStorage::DeleteTool(int number)
   {
-    if ((unsigned int)number > m_ToolCollection.size()) return false;
+    if (m_storageLocked)
+    {
+      MITK_WARN << "Storage is locked, cannot modify it!";
+      return false;
+    }
+
+    else if ((unsigned int)number > m_ToolCollection.size())
+    {
+      MITK_WARN << "Tool no " << number << "doesn't exist, can't delete it!";
+      return false;
+    }
     std::vector<mitk::NavigationTool::Pointer>::iterator it = m_ToolCollection.begin() + number;
     if(m_DataStorage.IsNotNull())
       m_DataStorage->Remove((*it)->GetDataNode());
@@ -85,13 +95,28 @@ bool mitk::NavigationToolStorage::DeleteTool(int number)
 
 bool mitk::NavigationToolStorage::DeleteAllTools()
   {
+   if (m_storageLocked)
+    {
+    MITK_WARN << "Storage is locked, cannot modify it!";
+    return false;
+    }
+
   while(m_ToolCollection.size() > 0) if (!DeleteTool(0)) return false;
   return true;
   }
 
 bool mitk::NavigationToolStorage::AddTool(mitk::NavigationTool::Pointer tool)
   {
-  if (GetTool(tool->GetIdentifier()).IsNotNull()) return false;
+  if (m_storageLocked)
+    {
+    MITK_WARN << "Storage is locked, cannot modify it!";
+    return false;
+    }
+  else if (GetTool(tool->GetIdentifier()).IsNotNull())
+    {
+    MITK_WARN << "Tool ID already exists in storage, can't add!";
+    return false;
+    }
   else
     {
     m_ToolCollection.push_back(tool);
@@ -129,4 +154,19 @@ int mitk::NavigationToolStorage::GetToolCount()
 bool mitk::NavigationToolStorage::isEmpty()
   {
   return m_ToolCollection.empty();
+  }
+
+void mitk::NavigationToolStorage::LockStorage()
+  {
+  m_storageLocked = true;
+  }
+
+void mitk::NavigationToolStorage::UnLockStorage()
+  {
+  m_storageLocked = false;
+  }
+
+bool mitk::NavigationToolStorage::IsLocked()
+  {
+  return m_storageLocked;
   }
